+++ app/mempool.go
@@ -1,31 +1,38 @@
 package app
 
 import (
-	"encoding/hex"
-	"fmt"
-	"github.com/ginchuco/ginchu/crypto"
-	"github.com/ginchuco/ginchu/state_machine"
 	"github.com/ginchuco/ginchu/state_machine/types"
 	lib "github.com/ginchuco/ginchu/types"
 )
 
 type Mempool struct {
-	mempool      lib.Mempool
-	mempoolStore lib.StoreI // ephemeral state to have safety within the mempool
-	mempoolState state_machine.StateMachine
+	State
+	lib.Mempool
+
+	log lib.Logger
 }
 
-// HandleTransaction handles mempool acceptance from incoming transactions
-// It alone decides if the transaction is valid based on the mempool state
-func (m *Mempool) HandleTransaction(tx, h []byte) lib.ErrorI {
-	if hash := hex.EncodeToString(h); m.mempool.Contains(hash) {
-		return types.ErrTxFoundInMempool(hash)
+func NewMempool(state State, log lib.Logger, config types.MempoolConfig) *Mempool {
+	return &Mempool{
+		Mempool: types.NewMempool(config),
+		State:   state,
+		log:     log,
 	}
-	fee, err := m.ApplyTransaction(tx)
+}
+
+// HandleTransaction accepts or rejects incoming txs based on the mempool state
+// - recheck when
+//   - mempool dropped some percent of the lowest fee txs
+//   - new tx has higher fee than the lowest
+//
+// - notes:
+//   - new tx added may also be evicted, this is expected behavior
+func (m *Mempool) HandleTransaction(tx []byte) lib.ErrorI {
+	fee, err := m.ApplyAndWriteTx(tx)
 	if err != nil {
 		return err
 	}
-	recheck, err := m.mempool.AddTransaction(tx, fee)
+	recheck, err := m.AddTransaction(tx, fee)
 	if err != nil {
 		return err
 	}
@@ -36,31 +43,35 @@ func (m *Mempool) HandleTransaction(tx, h []byte) lib.ErrorI {
 }
 
 func (m *Mempool) CheckMempool() lib.ErrorI {
-	if err := m.ResetMempoolState(); err != nil {
+	if err := m.ResetToBeginBlock(); err != nil {
 		return err
 	}
-	var toDelete [][]byte
-	it := m.mempool.Iterator()
-	defer it.Close()
-	for ; it.Valid(); it.Next() {
-		tx := it.Key()
-		if _, err := m.ApplyTransaction(tx); err != nil {
-			fmt.Println(err)
-			toDelete = append(toDelete, tx)
-		}
-	}
-	for _, tx := range toDelete {
-		if err := m.mempool.DeleteTransaction(tx); err != nil {
-			return err
-		}
+	var remove [][]byte
+	m.RecheckAll(func(tx []byte, err lib.ErrorI) {
+		m.log.Error(err.Error())
+		remove = append(remove, tx)
+	})
+	for _, tx := range remove {
+		m.DeleteTransaction(tx)
 	}
 	return nil
 }
 
-func (m *Mempool) ApplyTransaction(tx []byte) (fee string, err lib.ErrorI) {
+func (m *Mempool) RecheckAll(errorCallback func([]byte, lib.ErrorI)) {
+	it := m.Iterator()
+	defer it.Close()
+	for ; it.Valid(); it.Next() {
+		tx := it.Key()
+		if _, err := m.ApplyAndWriteTx(tx); err != nil {
+			errorCallback(tx, err)
+		}
+	}
+}
+
+func (m *Mempool) ApplyAndWriteTx(tx []byte) (fee string, err lib.ErrorI) {
 	txn, cleanup := m.TxnWrap()
 	defer cleanup()
-	result, err := m.mempoolState.ApplyTransaction(uint64(m.mempool.Size()), tx, crypto.HashString(tx))
+	result, err := m.ApplyTransaction(tx, m.Size())
 	if err != nil {
 		return "", err
 	}
@@ -69,19 +80,3 @@ func (m *Mempool) ApplyTransaction(tx []byte) (fee string, err lib.ErrorI) {
 	}
 	return result.Transaction.Fee, nil
 }
-
-func (m *Mempool) TxnWrap() (txn lib.StoreTxnI, cleanup func()) {
-	txn = m.mempoolStore.NewTxn()
-	m.mempoolState.SetStore(txn)
-	cleanup = func() {
-		m.mempoolState.SetStore(m.mempoolStore)
-		txn.Discard()
-	}
-	return
-}
-
-func (m *Mempool) ResetMempoolState() lib.ErrorI {
-	return m.mempoolStore.Reset()
-}
-
-func (m *Mempool) SetStore(store lib.StoreI) { m.mempoolStore = store }